add import from agent

+++ src/main.py
@@ -65,8 +65,7 @@ def import_pointcloud_ply(api: sly.Api, task_id: int):
                                  f"structure, for subdirectories and duplicated file names")
             continue
 
-    sly.fs.remove_dir(dir_=g.STORAGE_DIR)
-    if g.REMOVE_SOURCE:
+    if g.REMOVE_SOURCE and not g.IS_ON_AGENT:
         api.file.remove(team_id=g.TEAM_ID, path=g.INPUT_PATH)
         source_dir_name = g.INPUT_PATH.lstrip("/").rstrip("/")
         sly.logger.info(
